Project structure reorganization
